Delete data_helper.py: shifted to run_data_cleaning.py
